Float changed to int for time in Movie for less space used in db

+++ src/main/java/BE/Movie.java
@@ -52,7 +52,7 @@ public class Movie {
         return lastView;
     }
 
-    public float getTime() {
+    public int getTime() {
         return time;
     }
